Update Dockerfile (#266)
* Update Dockerfile

This is the Dockerfile we currently use to Build a docker-image for MAD.

As you see the dockerfile has two parts:
(1) In the first stage we have to compile tesseract, because the 3.7-slim python image has only access to tesseract 3.04 via apt. 
(2) In the second stage we copy the compiled binary and libaries  to our final image and then install all required packages and requirements to run MAD, finally we throw away everything we do not need to reduce image size even further.

We try to minimize the amount of layers, because the python-requirements and Open-CV blow up the image size drastically already. 
We end up with a final size of  916MB.

If you see (a) an opportunity to reduce the image size even further or (b) even come up with a solution to run it with an alpine base or (c) spot any bugs -  feel free to inform me :)

* pull tesseract from strech backport.

Worked in the proposal to pull tesseract from stretch-backport instead of a multistage build where we compile tesseract.
MAD still works as expected.

The image size even reduced to 765MB. (I found out this is because the compiled tesseract is 5.0 alpha, which contains alot of overhead)


-----------------------
Thanks @Breee