Fix upgrade issue around non-instanced devices
During the upgrade process for v19 it alters trs_status to use instance_id versus instance.  If the same device
exists twice (one with and one without an instance) it will cause an error during the upgrade.  This change
addresses that issue to notify it is removing non-instanced devices prior to altering the table