UX for route recalulation
During route recalcuation the UI now displays an icon showing that it is being being recalculated.  The 'areas' page now checks the status every 5 seconds to determine the status of all recalcuatlions fro all areas.  If the route is being recalculated it cannot be triggered again.

The schema for routecalcs has been updated to include the status (recalc_status) which states if the recalculation is running and when the last time the route was re-calculated.  These values are reset on boot since any previously running thread has been ended