add feature to host APK files through MAD (#659)
* Initial development of MADmin hosted APKs

This allows MAD to host any uploaded APKs.  This allows for local file-transfers when installing APKs for ATVs.

* APKs are now chunked when saving to the database

* downloading APKs now requires the same information present as the websocket connection

* Added Smart APK Pushing for jobs

Smart APK pushing validates architecture and version.  If MAD has a newer APK available it will install it to the device

* MADmin auto-search and download implementation

* a few fixes

* Whats the latest version of <insert apk here>?  Ask the wizard

Find and update all required APKs with the help of your friendly MADwizard!  Make sure to give them a poke on the MADmin APK page

* mysql prefers 255 vs 256 on indicies

* Dont really need that unique key

* Only push PoGo if the version is supported by addreses.json

* stating this is a beta feature