Update / Patcher refactoring (#707)
* Refactoring of patching

Patching is now handled on a per-patch basis versus one giant file for patching.  Each patch must pass validation for it to continue.  After a patch has been successfully applied, the version is updated so it will not run again

* Only mark the patch as completed after post-validation

* flake8 styling

* only state successful if it completed successfully

* Auto-install the latest DB schema

The patcher will attempt to install the latest DB schema if the MAD version cannot be found and the view is missing

* A few fixes so unit-tests pass on blank systems

* Fixes for the patcher

Testing, fixes, and validation for the patcher

* flake8 fixes

* Installations that set the schema now properly set the instance_id

* Cleanup the upgrade logic so its more readable

* removed 10 unused tables

* Remove the sql installation on travis  during boot as its now handled by the patcher