Several patcher fixes
If the current patch level was removed, the patcher was starting at the first patch level versus the next one
There is now a Primary Key check on the versions table.  If it does not exist, it will attempt to add it.  If there are duplicate keys in the table, it will de-dupe then apply the primary key