Patcher Fixes (#712)
* Several patcher fixes

If the current patch level was removed, the patcher was starting at the first patch level versus the next one
There is now a Primary Key check on the versions table.  If it does not exist, it will attempt to add it.  If there are duplicate keys in the table, it will de-dupe then apply the primary key

* Patch19 was not needed

Patch19 was redundant as Patch23 performed the same steps but had extra error handling.  Patch23 has been updated to work with both cases as 19 has been removed

* removed todo message related to unversioned changes

* Fixed an issue with an invalid variable reference

* Validate some trs_status fields

Sanity-check for v23 - v24 that the required trs_status fields exist

* check all elements that change in the trs_status table.  removed an unneeded unversioned change