prevent rare loop with ScreenType.ERROR/FAILURE
Had a device endlessly loop with ScreenType.ERROR/ScreenType.FAILURE,
endlessly logging "[ WARNING] Something wrong with screendetection
or pogo failure screen". It has to be assumed that
self._loginerrorcounter somehow was incremented over 2, thus MAD was
unable to enter the appropriate routine because it only checked == 2.
I changed this to check > 1 and to reset the loginerrorcounter to 0
once the routine ran.