PD APKs (#811)
* MAD APK available via PD for MAD ROM

The APKs are now accessible via the PD port so the madrom can easily pull the updated APKs

* validate auth before sending any data

* Origin Hopper

The Origin Hopper allows for an admin to rapidly deploy new devices through the PD receiver.  It can accept Walker and Pool IDs so no additional configure is required in MADmin.  If not present, it will use the first Walker and no Device Pool.