Remove configmode.py and move functionality to start.py (#840)
* Run start.py with config_mode flag

Updated start.py to run the same functionality as config_mode.  This will allow docker containers to easily run config_mode by updating the configuration files.  General cleanup of start.py and deprecation of config_mode.py

* Start MITM (PD) with config_mode

Starting MITM with config_mode allows MAD ROM (and other integrations) to always be able to pull the APKs from the server.  When booting in config_mode it will not accept any POSTs so devices should not attempt any work

* Fix for WorkerConfigmode

WorkerConfigmode was expecting a routemanager to be returned during its validation phase.  This is not applicable in config_mode which caused the box to constantly disconnect

* Moved two PD endpoints out of config_mode

These endpoints require MappingManager to be configured to retrieve data.  Since ConfigMode does not initialize this variable warning messages were appearing in the logs

* UnitTests can be run with configmode flag and skip any tests it cannot complete