handle login screen after wrong birthdate
when due to some sort of glitch a invalid birthdate is intered during
the login procedures (too young), we end up on a screen only showing PTC
and KIDS login buttons. It was found to be screentype.POGO which
resulted in an infinite loop. This commit properly identifies and
handles that situation.