Screendetection improvements (#850)
* detect ingame GPS errors again

due to false positives from heads up notifications, we ignored the upper
25% of the screen. Thus we also ignored red GPS errors in the game.
With this change, the upper 25% won't be ignored exclusively when checking
against the strings indicating GPS errors. This avoids the known false
positives from PogoDroid notifications while still being able to act on the
errors.

A slim chance for false positives remains, which could be handled by
disabling heads up notifications. They should be unnecessary on scanner
devices anyway.

* handle login screen after wrong birthdate

when due to some sort of glitch a invalid birthdate is intered during
the login procedures (too young), we end up on a screen only showing PTC
and KIDS login buttons. It was found to be screentype.POGO which
resulted in an infinite loop. This commit properly identifies and
handles that situation.

* use binarized pictures with different thresholds

... on screendetection.

On some screens, the new screentype.NOGGL being my test subject, the
text on some buttons couldn't be read by tesseract. By preprocessing the
screenshots, namely binarizing them with different thresholds, we can
improve results for text on different backgrounds.

* try pogo restart before reboot on frozen screen

as reboots are a potential cause for issues on ATV boxes and cause
additional wait time, it should be attempted to restart PoGo before
rebooting when the screen/game seems to be frozen.

* try to start PoGo when we find it closed

+ a little more logging

* minor screentype handling logic improvements