workaround: repeatedly click trash during questdel
there seems to be a recent issue with clicks on the trash quest icon not
registering. This leads to the procedure not properly emptying the quest
bag, which in turn reduces efficiency of the quest scan mode because
stops will be spun when the quest bag is already completely filled. This
causes the need to re-visit the stop later.

This change significantly reduces occurence of the problem (90%+
reduction) at the cost of some more delay for these clicks.

It's more of a brute-force workaround than a pretty fix though. More
research into the exact problem should yield a more efficient solution.