plugins: handle exceptions in walk_package, ignore hidden directories
* previously unhandled exceptions would stop loading subsequent plugins,
now it'll be visible which plugin caused issues and why, and be able to
continue with subsequent plugins

* recursing into hidden directories caused an exception because the
contruction of package names as (package + '.' + child_pkg) with
child_pkg starting with a dot seems to have caused invalid package
names. There should be no reason to recurse into directories that are
hidden / starting with a dot.