Fix route manager locking
'with a and b:' does not do what one might think. The correct
syntax is 'with a, b:' or 'with a as foo, b as bar:', etc.

It seems that 'with a and b:' does require that 'a' have an
__enter__ method, but it never calls it... it just skips to 'b'
as can be seen below.

Additionally, had both locks worked, there was a lock ordering problem
in one spot.

```
class Testing:
    def __enter__(self):
        print('    enter: %s' % self.__class__)
        return 1
    def __exit__(self, a, b, c):
        print('    exit: %s' % self.__class__)

class Testing1(Testing):
    pass

class Testing2(Testing):
    pass

t1 = Testing1()
t2 = Testing2()

print('testing with "and":')
with t1 and t2:
    print('woot')

print()

print('testing with comma:')
with t1, t2:
    print('woot')

----- OUTPUT ------

testing with "and":
    enter: <class '__main__.Testing2'>
woot
    exit: <class '__main__.Testing2'>

testing with comma:
    enter: <class '__main__.Testing1'>
    enter: <class '__main__.Testing2'>
woot
    exit: <class '__main__.Testing2'>
    exit: <class '__main__.Testing1'>
```