Fix dropped prioq coord for iv_mitm
When multiple workers are working on the same iv_mitm area, it
appears that prioq coords can be passed between workers, but
they are never looked for:

```
    elif self._routepool[origin].has_prio_event and self.mode != 'iv_mitm':
          prioevent = self._routepool[origin].prio_coords
          route_logger.info('getting a nearby prio event {}', prioevent)
          [...]
```

Additionally, it appears some sort of stall of the route manager
can occur when that happens, but I am not quite sure the cause
of that.

So, I suppose either that 'iv_mitm' check above could be removed, or we
could just not attempt to pass coords to another worker when we're in iv_mitm
mode. I'm choosing the latter, because the above 'elif' makes me believe
it never should have passed one in the first place.

Also: Add some safety to ensure we enter the block of code for checking
the prioq when mode is 'iv_mitm'... and recursing when we find we have
raced with another worker and nothing is in queue.