Fix websocket/worker races
Often in our setup, RGC reconnects before MAD notices that the old
connection is dead. When this happens, the worker is found to still
be running and not shutting down... thus the new connection is just
replaced in the WebsocketConnectedClientEntry for the worker to use.

However, once MAD finally detects the old connection is dead, it then
stops the worker... leaving no worker running for the new RGC
connection. The new RGC connection continues to ping/pong just fine,
but it's doing nothing else.. You have to restart the device (or MAD)
to recover from this.

This fixes the issue by moving the stopping of the worker to the
connection handler and only doing it when the 'Done' connection is
still the one in the WebsocketConnectedClientEntry.

But we also need to not allow new connections while the worker is still
running on an old one because of potential races with the worker
shutting itself down.

And also: don't replace the websocket connection in the Entry if we're
not registering it.

Also modifies logs, which helped track this down and may be useful
in the future.