Parse routes in legacy MAD
This code is backported from async. To receive Route protos, you
must be running the latest legacy PD.

I did not create a DB migration for this. Just paste the table
creation that you see in the code into mysql/mariadb.

I have not tested this exact code, as I run a modified version of
legacy MAD. But I'm pretty confident this is fine.